Create a common FakeLifecycle and LifecycleOwner (#1471)
